[GC] [VC] Remove the use of wasm imported types and generated contract clients for cross contract interactions.
Sharing types in cross contract interactions with contractimport! causes a false sense of interoperability, as they work in unit tests but not in futurenet. The error was:

thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: MissingEntry("ProposalPayload")'

The solution reached here was to:
1. Create a third crate "shared" that will hold the shared structures per contract until we find a better solution.
2. Remove the use of contract clients for cross contract interactions. Now we need to convert all the types to `Val` before interacting with the dependent contract and sending them through env.invoke_contract(...)
