+++ Cargo.toml
@@ -21,6 +21,7 @@ resolver = "2"
 # Our workspace members include the packages in the contracts directory.
 members = [
     "contracts/governance",
+    "contracts/shared",
     "contracts/voting",
     "contracts/simple_voting",
     "contracts/raffle",
