+++ contracts/governance/Cargo.toml
@@ -10,6 +10,7 @@ crate-type = ["cdylib"]
 
 [dependencies]
 soroban-sdk = { workspace = true }
+shared = { path= "../shared"}
 
 [dev-dependencies]
 soroban-sdk = { workspace = true, features = ["testutils"] }
