+++ contracts/shared/Cargo.toml
@@ -1,15 +1,12 @@
 [package]
-name = "voting"
+name = "shared"
 version = "0.1.0"
 edition = "2021"
 
-[lib]
-crate-type = ["cdylib"]
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 soroban-sdk = { workspace = true }
-shared = { path = "../shared" }
 
 [dev-dependencies]
 soroban-sdk = { workspace = true, features = ["testutils"] }
-rstest = "0.17.0"
